@@ -12,27 +12,26 @@
   },
   "dependencies": {
     "dotenv": "^16.0.2",
-    "eslint-config-prettier": "^8.5.0",
-    "prettier": "^2.7.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@testing-library/react": "^13.3.0",
-    "@types/jest": "^28.1.8",
+    "@types/jest": "^29.4.0",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
-    "@vitejs/plugin-react": "^2.0.1",
+    "@vitejs/plugin-react": "^3.1.0",
     "eslint": "^8.23.0",
-    "jest": "^28",
+    "eslint-config-prettier": "^8.6.0",
+    "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.0.1",
     "lefthook": "^1.1.1",
-    "prettier": "^2.7.1",
-    "ts-jest": "^28.0.8",
-    "typescript": "^4.8.2",
-    "vite": "^3.0.7"
+    "prettier": "^2.8.3",
+    "ts-jest": "^29.0.5",
+    "typescript": "^4.9.5",
+    "vite": "^4.1.1"
   },
   "license": "MIT"
 }
