@@ -14,26 +14,27 @@
     "cors": "^2.8.5",
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.2",
-    "eslint-config-prettier": "^8.5.0",
     "express": "^4.18.1",
-    "pg": "^8.8.0",
-    "prettier": "^2.7.1"
+    "pg": "^8.8.0"
   },
   "devDependencies": {
+    "@babel/core": ">=7.8.0 <8.0.0",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
-    "@types/jest": "^28.1.8",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.11.18",
     "@types/pg": "^8.6.5",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
-    "babel-jest": "^29.0.1",
-    "eslint": "^8.23.0",
-    "jest": "^28",
-    "nodemon": "^2.0.19",
-    "prettier": "^2.7.1",
-    "ts-jest": "^28.0.8",
+    "babel-jest": "^29.4.1",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
+    "jest": "^29.4.1",
+    "nodemon": "^2.0.20",
+    "prettier": "^2.8.3",
+    "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
-    "typescript": "^4.8.2"
+    "typescript": "^4.9.5"
   },
   "license": "MIT"
 }
